+++ apps/web/next.config.mjs
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  // output: "exports",
+  output: "export",
 
   images: {
     remotePatterns: [
